scrape/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-pubsub==0.42.1
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.11.4
grpcio==1.22.0
idna==2.8
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
